Abstract to conbcrete syntax conversion step 3.
Moved logic to calculate the amount of times a production should be
repeated when converting frequency modifiers to an exteral service, so
that it could be reused for the third step.
Added HasProduction interface to accomunate nodes holding productions
and allowing them to be swapped for new productions (which is what we
need to do in this step).