Added static syntax check.
Added static syntax check, verifying that non termianting symbols always
have a declaration in the pertaining scope and that the start symbol is
always defined.
I can always go back and add warnings, but for now it's enough to just
check for errors.
Made up my mind about "declaration" versus "definition" naming. In case
I forget, "declaration" is the generic term used to refer to both
Definitions and Assignments.
Expanded AbstractSyntaxWalker interface to allow for a second object to
be passed as context to the walker.
This should also enable the last abstract to concrete syntax conversion
step.