Infinite recursion static check.
Halfway through the implementation of this I noticed that Polygen does
not warn or error out when an infinite recursion loop is found in the
declarations, despite the documentation stating otherwise and even when
using the -pedantic flag (an alternative explanation is that maybe my
binary is outdated?).
But I was in the middle of it, so I decided to finish the implementation
anyway.
I thought about adding the mentioned warnings when a potential infinite
loop is encountered (one where some of the possible outputs of a
declaration triggers an infinite loop, but some others don't), but in
the end I decided to not do that, at least for now.
I want to get this thing working, there will be time for optimizations
afterwards.