Concrete to abstract syntax conversion, last step.
With this commit the concrete to abstract syntax conversion has
been finally implemented.
Excluding bugs, it should be possible to convert a grammar in abstract
syntax into one with concrete syntax.
I noticed that the conversions operated by the last step and the one
before it are basically the same: they both take an unfolded unfoldable
and surface its production to where the unfoldable is unfolded. So I
captured the shared logic into an abstract parent class and had both
converters extend them and only provide minimal methods to extract
productions and declarations from the unfoldable to unfold.