ELD-99: Imroved blank lines 24.67x on my 1337 933|< particle accelerator.
This commit is, if possible, even less refined than the ones preceeding
it. My unstated challenge for these quarantine days was to take back
this project started one day in 2016, and bring it to completion.
Tomorrow, the most restrictive quarantine lockdown rules will be
released for the first time in Italy, so I wanted to make sure to get
this done by tomorrow.
In order to do that, something had to go. This commit does not have any
test, but hopefully I will be able to add them in the coming days, as
the stricter lockdown is lifted, but the quarantine is not over yet.
The commit message is the output of a run of polygen-php against the
test grammar that you can find in tests/files folder.
The second run ended up in an infinite loop, which tells me that there
is something wrong either in one of the conversion steps or in the
generation / interpretation phase.