Bugfixes, refactoring and new features.
Renamed Context to DeclarationCollection and moved to Utils namespace.
Removed ContextInterface as it was serving no purpose.
Made interpretation start symbol customizable, and updated no start
symbol error and static check to accept the start symbol name as well.
Added ProductionCollection class to hold a set of productions and moved
there the logic to select a production at random and to return a new
ProductionCollection given a label selection.
Added label selection support to interpretation context, to allow nested
contexts to inherit partent contexts' selected labels.
Reworked assignments to be tied to the DeclarationContext the are
declared in, allowing nested contexts to re-declare the assignment if
wanted.
Added a skipped test to support consecutive label selections that are
mentioned in the manual, supported in the actual client, but not
represented in the grammar definitions of the concrete syntax.
Allowed retrieval of seed used in context, if no seed is specified
(used to debug interpretation errors).
Made "\n" token always output PHP_EOL instead of "\n". This should also
be configurable, since line endings can be chosen in the polygen
command line options, but it will be a future addition, if necessary.
For now, being consistent with the platform on which it runs is more
desirable and easily achievable.