Fixed missing dependency.
Added webmozart/assert as dependency.
Made start symbol optional from the Polygen->getDocument() method. I
don't even particularly like that it is a parameter for that method, but
I also want the static syntax check to be done before returning the
document, rather than when processing it, so for now, it must be passed
twice (but it's optional in both places for the general case).