Added support for consecutive label selections.
Got rid of the "HasLabelSelection" interface: only Atom was implementing
it.
Added LabelSelectionCollection and reworked some of the logic to work
with them instead that plain LabelSelection.
Updated parser to allow consecutive label selections, label selection
groups and label selection reset indicators.
Added an extra bonus conversion step to deal with this sequence.